Fix the way to get BUILD number for WSL images
The way it is parsing the build number now generates 2 numbers if
the string contains `SnapshotX` and `BuildY`.
It even returns 2 numbers and messes up the openqa call.

e.g.
SUSE-Linux-Enterprise-Server-15-SP3--Snapshot12--x64-Build2.209.appx
returns
    12
    2.209
and it keeps BUILD=12 and 2.209 at the end of the call without variable
assignation, which makes the call to fail.

With this fix, we make sure we keep only 1 of the numbers, specially
the one at the end.