Schedule product per livepatch kernel_version
Automate the livepatch isos posts. Due to the logic of the script generation
this commit inserts a for loop for all the livepatches which
will be scrapped from test repository.

- the loop always run even if no livepatches found. However it will only
append to the isos posts the variables only in the case that actually livepatch
versions are available. To get the delta of the increment repo, it checks the
datetime.
- Use the same uri which is used for rsync. For that, some manipulation is
needed to convert the rsync to a regular http URI.

The approach is just a bit dirty, as any other approach would require further
refactoring and bigger changes.

depends-on: https://gitlab.suse.de/openqa/openqa-trigger-from-ibs-plugin/-/merge_requests/263
issue: https://progress.opensuse.org/issues/190134
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>